[API] Split integration tests that use different models
