+++ aws-amplify-api-aws/src/androidTest/res/raw/amplifyconfiguration.json
@@ -35,12 +35,13 @@
     "plugins": {
       "AWSAPIPlugin": {
         "GraphQLInstrumentationTest": {
-          "Endpoint": "https://bb2xm5ttqvbfdeu5oi5gw7cvma.appsync-api.us-east-1.amazonaws.com/graphql",
+          "Endpoint": "https://57qnjcavhbdatfveg5dccboqji.appsync-api.us-west-2.amazonaws.com/graphql",
           "Region": "us-east-1",
           "AuthorizationType": "API_KEY",
-          "ApiKey": "da2-og4mnlou65dm5d6hs2itesgkzm"
+          "ApiKey": "da2-2gvk5ltfync4linsd2mhxx35my"
         }
       }
     }
   }
 }
+
